Initial implementation of custom keys.
This feature allows to customize the type of a logical key in the
keyboard, allowing the user to perform any arbitrary action when
a key changes the state in the matrix. This is done by implementing
the HandleKey trait over a user defined type, that will determine
which action will be ran with each key.

It is also responsability of the HandleKey to add or remove
the pressed keys from the HID report that will be sent out to
the host. This opens the possibility to arbitrarily map
logical keyboard keys from zero or more HID keys.

The original LayoutKey implementation that used to represent
the way a Key was defined in the keyboard becomes "DefaultKey", that
implements HandleKey, as a standard provided implementation of a key
for those users that don't want any further customization of their
keys.