Added layers! proc macro support for arbitrary aliases via macro_rules.
With the new implementation of custom key type definition,
the layers! proc macro becomes less useful if it doesn't
support custom key definitions as well.

This commits makes more flexible the definition of aliases
for keys by removing the hardcoded keymap.rs from the
dxkb-proc-macro, and delegating the mappings between the aliases
and the keys to a custom user defined proc macro. The proc-macro
will just print a call to the macro with the received tokens from
the layers definition, and the macro rules will do the rest.

The commit includes an implementation for this macro_rules for the
DefaultKey, and a base macro, hid_key_from_alias, that defines
a set of default aliases between standard HID keyboard keys and their
KeyboardUsage, and to which other custom defined alias resolvers can
delegate into.

Still, the path of the alias macro is hardcoded and pointing to the
default_key_from_alias!. That will be allowed in a following-up
commit.