Added alias_resolver field to the layers proc macro
This field allows users to specify a macro_rules path that will
be used for resolving aliases of the keys. If no specified,
then the default (dxkb_core::default_key_from_alias), will be used
and therefore it will use the default key type. Check the definition
of this macro for seeing an example of an implementation of this
kind of macros.